Move transform_manager to submodule
