+++ pytransform3d/transform_manager/__init__.py
@@ -1,0 +1,9 @@
+"""Manage complex chains of transformations.
+
+See :doc:`transform_manager` for more information.
+"""
+from ._transform_graph_base import TransformGraphBase
+from ._transform_manager import TransformManager
+
+
+__all__ = ["TransformGraphBase", "TransformManager"]
